demonstrating ideas how to separate BCD from BCD Signalling
