+++ webpack.config.js
@@ -73,6 +73,15 @@ module.exports = [
         ...commonConfig
     },
     {
+        target: 'web',
+        entry: path.resolve(__dirname, './kuma/javascript/src/bcd-signal.jsx'),
+        output: {
+            filename: 'bcd-signal.js',
+            path: path.resolve(__dirname, './kuma/javascript/dist/')
+        },
+        ...commonConfig
+    },
+    {
         target: 'node',
         entry: path.resolve(__dirname, './kuma/javascript/src/ssr.jsx'),
         output: {
